// frontend/src/api/generated/models/fusionDashboardDto.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * 在线答题与智能测评平台 API
 * Online exam and intelligent assessment platform API
 * OpenAPI spec version: 1.0
 */
import type { DashboardDrilldownDto } from './dashboardDrilldownDto';
import type { FusionAcademicMetricsDto } from './fusionAcademicMetricsDto';
import type { FusionAssessmentMetricsDto } from './fusionAssessmentMetricsDto';
import type { FusionDashboardDtoRole } from './fusionDashboardDtoRole';
import type { FusionRecentExamDto } from './fusionRecentExamDto';
import type { TeacherPerformanceDto } from './teacherPerformanceDto';

export interface FusionDashboardDto {
  academic: FusionAcademicMetricsDto;
  assessment: FusionAssessmentMetricsDto;
  drilldowns: DashboardDrilldownDto[];
  from: string;
  recentExams: FusionRecentExamDto[];
  role: FusionDashboardDtoRole;
  scopeLabel: string;
  teacherPerformance: TeacherPerformanceDto[];
  to: string;
}